[cmd/health] Do not resolve secrets + fix CLI options (#7149)
* [cmd/health] Do not decrypt secrets in health command

The health command is called from the container liveness probe script,
which is called regularly and frequently. Avoid decrypting the secrets
when this command is called since decrypting secrets can be a costly
operation.

* [cmd/health] Fix command's CLI arguments

They were not being passed properly by reference to the command, and
therefore they were being ignored.