Re-order offset guessing of dport and family fields (#7160)
This has the `family` guessing performed after `dport`. This is because they are in that order in the `__sk_common` struct. We also start the family guessing at the `dport` offset. This is because there is a hash between `saddr` and `dport`, which has a high likelihood of falsely matching `family`.