[gitlab] Allow docker release jobs to target dev repositories on non-tagged pipelines (#7205)
Allows docker release jobs to release images to the agent-dev repo on non-tagged deploy pipelines.

Refactors the release jobs: uses a common script for Dockerhub and GCR jobs, splits the dogstatsd release jobs.

First part of a 2-part refactor of the Gitlab jobs used to release the docker images. The second part will move these jobs to dedicated child pipelines.

Currently, in each job, there is a bash / powershell line which adds -dev to the repository name if we don't follow the deploy_on_tag_* rules (so that only deploy_on_tag_* pipelines can release to prod repositories). This is a temporary workaround: the long-term solution would be to use Gitlab rules directly to set variables, but that feature got added in Gitlab 13.8, which we don't use yet.