[pkg/collector/check] Initialize check fail/success telemetry counters (#7243)
* [pkg/telemetry] Allow Promethius counters to be initialized

This change adds a method to initialize a prometheus counter (to 0).
This change should enable us to provide telemetry about counters which
are important to track even when they are 0 and before they are
incremented (ex. check failure/success counts).

* [pkg/telemetry] Allow our telemetry counters to be initialized

This change adds a method to initialize out telemetry counters (to 0)
via the API. This change should enable us to provide telemetry about
counters which are important to track even when they are 0 and before
they are incremented (ex. check failure/success counts).

* [pkg/collector/check] Initialize check fail/success telemetry counters to 0

When telemetry is turned on, we previously were unable to ascertain the
number of fail/ok checks since only one of the tags (`"state":"ok"` or
`"state":"fail"`) would be set on each check. If a particular check had
no failures (or successes), the telemetry for the opposite state (of
value 0) would not be sent and aggregation of data would not be accurate.

This change ensures that when a check is scheduled we initialize both
the "ok" and "fail" counts to 0 which should fix the aggrewgation.

Upstream bug: https://datadoghq.atlassian.net/browse/AC-657